refactor: keep execution tests within size ratchet
