+++ pkg/services/factory_sessions/internal/execution/javascript_runtime_persistence_test.go
@@ -729,54 +729,6 @@ func TestPersistenceChoiceForPolicy_EnabledCreatesProjectDurableSessions(t *test
 	}
 }
 
-func exactEncodedSizeWarningState(t *testing.T, targetSize int) runtimeSessionState {
-	t.Helper()
-	state := runtimeSessionState{
-		session: SessionReadResult{
-			SessionID: "dur-sess-warning-threshold",
-			Status:    LifecycleStatusSucceeded,
-		},
-		petriMutations: []interfaces.TokenMutationRecord{{
-			Type:                interfaces.MutationCreate,
-			TokenID:             "live-token",
-			ToPlace:             "task:running",
-			TransitionReachable: true,
-			Token: &workers.Token{
-				ID:    "live-token",
-				Color: workers.Color{WorkID: "live-work"},
-			},
-		}},
-		petriSummaries: []PetriTokenSummary{{
-			TokenID: "terminal-token",
-			WorkID:  "terminal-work",
-			PlaceID: "task:done",
-		}},
-	}
-	base := encodedWarningStateBytes(t, state)
-	state.sourceContent = "x"
-	perByte := encodedWarningStateBytes(t, state) - base
-	if perByte != 1 {
-		t.Fatalf("sourceContent encoded byte delta = %d, want 1", perByte)
-	}
-	if targetSize < base {
-		t.Fatalf("target snapshot size %d is below base size %d", targetSize, base)
-	}
-	state.sourceContent = strings.Repeat("x", targetSize-base)
-	if got := encodedWarningStateBytes(t, state); got != targetSize {
-		t.Fatalf("constructed snapshot bytes = %d, want %d", got, targetSize)
-	}
-	return state
-}
-
-func encodedWarningStateBytes(t *testing.T, state runtimeSessionState) int {
-	t.Helper()
-	encoded, err := json.MarshalIndent(persistedSnapshotFromRuntimeStateWithFailureLogCapacity(state, 0), "", "  ")
-	if err != nil {
-		t.Fatalf("marshal warning state: %v", err)
-	}
-	return len(encoded)
-}
-
 func TestPrepareStartAndPersistenceHelpers(t *testing.T) {
 	t.Parallel()
 	projectRoot := writeSimpleFinalWorkflowProject(t)
