+++ pkg/services/factory_sessions/internal/execution/javascript_runtime_start_test.go
@@ -815,7 +815,6 @@ func scriptedFailedRuntimeWorkflows() factory.JavaScriptWorkflows {
 		}, nil
 	})
 }
-
 func inlineWorkflowStartRequest(
 	requestID string,
 	source string,
@@ -839,7 +838,6 @@ func inlineWorkflowStartRequest(
 		RequestedPolicy: requestedPolicy,
 	}
 }
-
 func waitUntilSessionStatus(
 	t *testing.T,
 	service Service,
@@ -866,7 +864,6 @@ func waitUntilSessionStatus(
 	t.Fatalf("session %s did not reach status %q within %s", sessionID, want, timeout)
 	return SessionReadResult{}
 }
-
 func decodePrimaryResultMap(t *testing.T, raw json.RawMessage) map[string]any {
 	t.Helper()
 
@@ -889,7 +886,6 @@ func decodePrimaryResultMap(t *testing.T, raw json.RawMessage) map[string]any {
 	t.Fatalf("primary result content = %#v, want JSON part", content)
 	return nil
 }
-
 func writeSimpleFinalWorkflowProject(t *testing.T) string {
 	t.Helper()
 
@@ -915,7 +911,6 @@ func orchestrationJavaScriptFromWorkflows(workflows factory.JavaScriptWorkflows)
 	}
 	return orchestrationJavaScriptAdapter{workflows}
 }
-
 func (a orchestrationJavaScriptAdapter) RunJavaScript(
 	ctx context.Context,
 	req factory.JavaScriptRuntimeRequest,
@@ -930,7 +925,6 @@ func (a orchestrationJavaScriptAdapter) ResumeJavaScript(
 ) factory.JavaScriptResumeContext {
 	return a.ResumeContext(summary, records)
 }
-
 func newTerminalWorkersService(t *testing.T, provider providers.Service) WorkerExecution {
 	t.Helper()
 	return terminalWorkerService{provider: provider}
@@ -982,4 +976,24 @@ func (service terminalWorkerService) Execute(
 	result.Outcome = workers.ExecutionOutcomeAccepted
 	result.Output.Primary = []work.WorkContentPart{{Text: providerResult.Content}}
 	return result, nil
+}
+
+type petriCompactionStore struct {
+	saveCalls int
+	saveErr   error
+	payload   []byte
+}
+func (s *petriCompactionStore) Save(_ string, encoded []byte) error {
+	s.saveCalls++
+	if s.saveErr != nil {
+		return s.saveErr
+	}
+	s.payload = append([]byte(nil), encoded...)
+	return nil
+}
+func (s *petriCompactionStore) Load(string) ([]byte, error) {
+	if len(s.payload) == 0 {
+		return nil, errors.New("not found")
+	}
+	return append([]byte(nil), s.payload...), nil
 }
